+++ libs/model/src/index.ts
@@ -2,18 +2,15 @@ export * from './lib/model.module';
 export * from './lib/address.model';
 export * from './lib/address.save.args.model';
 export * from './lib/address.service';
-export * from './lib/address';
 export * from './lib/base.entity.service';
 export * from './lib/base.model';
 export * from './lib/base.save.args.model';
 export * from './lib/country.model';
 export * from './lib/country.save.args.model';
 export * from './lib/country.service';
-export * from './lib/country';
 export * from './lib/entity.base';
 export * from './lib/model.module';
 export * from './lib/organization.model';
 export * from './lib/organization.save.args.model';
 export * from './lib/organization.service';
-export * from './lib/organization';
-export * from './lib/unique.display.entity.base';
+
