+++ libs/model/src/lib/bank.service.ts
@@ -1,8 +1,7 @@
-import { BaseEntityService } from '@erp2/model';
+import { Bank, BaseEntityService } from '@erp2/model';
 import { BankModel } from './bank.model';
 import { BankSaveArgsModel } from './bank.save.args.model';
 import { EntityManager, Repository } from 'typeorm/index';
-import { Bank } from './bank';
 
 export const BankServiceKey = 'BankService';
